Simplify datamodel-connector API
We can use Diagnostics and parser-database in there now.
